requirements.txt
tensorflow==2.17.0
keras==3.4.1
numpy== 1.26.4
pandas== 2.2.2
opencv-python==4.10.0.84
matplotlib==3.7.1
streamlit==1.38.0
Pillow==10.3.0